+++ demo/src/components/controls/number.tsx
@@ -41,16 +41,18 @@ export default function NumberControlView() {
         <wup-num
           w-name="number"
           w-label="Number"
-          w-initValue="1234"
+          w-initValue={1234}
           w-validations="window.myNumberValidations"
           w-mask=""
           w-maskholder=""
           w-prefix=""
           w-postfix=""
+          w-scale={1}
+          w-offset={0}
         />
         <div className={stylesCom.group}>
-          <wup-num w-name="readonly" readonly w-initValue="123" />
-          <wup-num w-name="disabled" disabled w-initValue="123" />
+          <wup-num w-name="readonly" readonly w-initValue={123} />
+          <wup-num w-name="disabled" disabled w-initValue={123} />
           <wup-num w-name="required" w-validations="myRequiredValidations" />
         </div>
         <wup-num
@@ -72,6 +74,54 @@ export default function NumberControlView() {
           w-storageKey
           w-storage="url"
         />
+        <div className={stylesCom.group}>
+          <wup-num
+            w-name="scaled"
+            w-label="With option scale 0.01"
+            w-initValue={0.53}
+            w-scale={0.01}
+            ref={(el) => {
+              if (el) {
+                const el2 = el.nextElementSibling as WUPNumberControl;
+                setTimeout(() => (el2.$value = el.$value), 10);
+                el.$onChange = () => (el2.$value = el.$value);
+              }
+            }}
+          />
+          <wup-num
+            w-name="scaledReal"
+            w-label="Real value"
+            ref={(el) => {
+              if (el) {
+                el.$options.format = { maxDecimal: 30 };
+              }
+            }}
+          />
+        </div>
+        <div className={stylesCom.group}>
+          <wup-num
+            w-name="shifted"
+            w-label="With option offset 20"
+            w-initValue={1234}
+            w-offset={20}
+            ref={(el) => {
+              if (el) {
+                const el2 = el.nextElementSibling as WUPNumberControl;
+                setTimeout(() => (el2.$value = el.$value), 10);
+                el.$onChange = () => (el2.$value = el.$value);
+              }
+            }}
+          />
+          <wup-num
+            w-name="shiftedReal"
+            w-label="Real value"
+            ref={(el) => {
+              if (el) {
+                el.$options.format = { maxDecimal: 30 };
+              }
+            }}
+          />
+        </div>
         <button type="submit">Submit</button>
       </wup-form>
     </Page>